+++ src/Sentry.Unity/Properties/AssemblyInfo.cs
@@ -2,5 +2,6 @@ using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Sentry.Unity.Tests")]
 [assembly: InternalsVisibleTo("Sentry.Unity.Editor")]
+[assembly: InternalsVisibleTo("Sentry.Unity.iOS.Tests")]
 [assembly: InternalsVisibleTo("Sentry.Unity.Editor.iOS")]
 [assembly: InternalsVisibleTo("Sentry.Unity.Editor.Tests")]